Production smoke tests

+++ api/osf_api.py
@@ -16,6 +16,9 @@ def current_user(session):
     user.get()
     return user
 
+def get_preferred_node(session):
+    return client.Node(session=session, id=settings.PREFERRED_NODE)
+
 def get_user_institutions(session, user=None):
     if not user:
         user = current_user(session)
@@ -35,14 +38,18 @@ def get_all_institutions(session):
     return institutions
 
 def delete_all_user_projects(session, user=None):
+    """Delete all of your user's projects that they have permission to delete
+    except PREFERRED_NODE (if it's set).
+    """
     if not user:
         user = current_user(session)
     nodes_url = user.relationships.nodes['links']['related']['href']
     data = session.get(nodes_url)
     for node in data['data']:
-        n = client.Node(id=node['id'], session=session)
-        n.get()
-        n.delete()
+        if node['id'] != settings.PREFERRED_NODE:
+            n = client.Node(id=node['id'], session=session)
+            n.get()
+            n.delete()
 
 def waffled_pages(session):
     waffle_list = []
@@ -53,7 +60,25 @@ def waffled_pages(session):
             waffle_list.append(page['attributes']['name'])
     return waffle_list
 
+def get_existing_file(session, node_id=settings.PREFERRED_NODE):
+    """Return the name of the first file in OSFStorage on a given node.
+    Uploads a new file if one does not exist.
+    """
+    node = client.Node(session=session, id=node_id)
+    node.get()
+    files_url = node.relationships.files['links']['related']['href']
+    data = session.get(files_url + 'osfstorage/')
+    file = data['data']
+    if file:
+        return data['data'][0]['attributes']['name']
+    else:
+        return upload_fake_file(session, node)
+
 def upload_fake_file(session, node, name='osf selenium test file for testing because its fake.txt'):
-    # Note: I gave this file a long name because it makes it easier to click if it takes up more space
-    session.put(url='{}/v1/resources/{}/providers/osfstorage/'.format(settings.FILE_DOMAIN, node.id), query_parameters={'kind': 'file', 'name': name}, raw_body={})
+    """Upload an almost empty file to the given node. Return the file's name.
+
+    Note: The default file has a very long name because it makes it easier to click a link to it.
+    """
+    session.put(url='{}/v1/resources/{}/providers/osfstorage/'.format(settings.FILE_DOMAIN, node.id),
+                    query_parameters={'kind': 'file', 'name': name}, raw_body={})
     return name
